Moved constants and endpoints to utils folder
