+++ sdk/interfaces/__tests__/login.test.js
@@ -1,6 +1,6 @@
 import { login } from '../index';
 import makeRequest from '../../requests';
-import REQUEST_TYPES from '../../requests/constants';
+import REQUEST_TYPES from '../../utils/constants';
 
 jest.mock('../../requests');
