+++ sdk/interfaces/index.js
@@ -1,5 +1,5 @@
 import makeRequest from '../requests';
-import REQUEST_TYPES from '../requests/constants';
+import REQUEST_TYPES from '../utils/constants';
 import { setParameters } from '../configuration';
 
 const initialize = (redirectUri, clientId, clientSecret) => {
