+++ sdk/utils/constants.js
